rename keyboard nav hook file
